Fixat filmCategory och read och delete och lite annat smått

+++ src/main/java/Business/Category.java
@@ -9,7 +9,7 @@ public class Category {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
     @Column(name = "category_id")
-    private int categoryId;
+    private byte categoryId;
     @Basic
     @Column(name = "name")
     private String name;
@@ -31,13 +31,11 @@ public class Category {
 
     }
 
-    public int getCategoryId() {
+    public byte getCategoryId() {
         return categoryId;
     }
 
-    public void setCategoryId(int categoryId) {
-        this.categoryId = categoryId;
-    }
+    public void setCategoryId(byte categoryId) {this.categoryId = categoryId;}
 
     public String getName() {
         return name;
